demos/pyairbyte_mindsb/requirements.txt
python-dotenv
pytest
psycopg2-binary
SQLAlchemy
airbyte
mysql-connector-python
